better batching

+++ index.js
@@ -3,18 +3,30 @@ const { randomPoint } = require('@turf/random');
 
 const outFile = 'points.csv';
 const numPoints = 30000000;
+const batch = 1000000;
 
 const bbox = [-180, -90, 180, 90];
 
-fs.writeFileSync(outFile, 'id,longitude,latitude\n', 'utf-8');
+// Write headers if our file doesn't exist yet, aka we're running this for the first time
+if (!fs.existsSync(outFile)) {
+	fs.writeFileSync(outFile, 'id,longitude,latitude\n', 'utf-8');
+}
 
-for (let i = 0; i < numPoints; i++) {
-	const point = randomPoint(1, bbox);
-	const coords = point.features[0].geometry.coordinates;
-	const file = `id-${i},${coords[0]},${coords[1]}`;
+let rowId = 0;
+for (let i = 0; i < (numPoints / batch); i++) {
+	const points = randomPoint(batch, bbox);
+	const lines = points.features.map(feature => {
+		const coords = feature.geometry.coordinates;
+		const line = `id-${rowId++},${trimCoord(coords[0])},${trimCoord(coords[1])}`;
+		if (rowId % 1000000 === 0) {
+			console.log('Processed', i * batch, 'points');
+		}
+		return line;
+	}).join('\n');
 
-	fs.appendFileSync(outFile, file + '\n', 'utf-8');
-	if (i % 100000 === 0) {
-		console.log('Processed', i, 'points');
-	}
+	fs.appendFileSync(outFile, lines + '\n', 'utf-8');
+}
+
+function trimCoord (coord) {
+	return coord.toFixed(2);
 }
